requirements.txt
# Spotify ETL Project Dependencies
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1

# Для тестирования
#pytest==8.3.2


#pip freeze > requirements.txt
